# Conflicts:
#	GameProject/Project/GameProject/Game/Game.cpp

+++ GameProject/Project/GameProject/Game/GameData.h
@@ -1,5 +1,5 @@
 #pragma once
 
 class GameData {
-	static int km;
+	static float km;
 };
